Add docstrings and refactor code
