+++ src/utils/forms.py
@@ -1,0 +1,30 @@
+"""
+Module for forms
+"""
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, SubmitField
+from wtforms.validators import DataRequired, Email, EqualTo
+
+
+class UserForm(FlaskForm):
+    """User form for registration"""
+
+    name = StringField("Name", validators=[DataRequired()])
+    email = StringField("Email", validators=[DataRequired(), Email()])
+    password = PasswordField(
+        "Password",
+        validators=[
+            DataRequired(),
+            EqualTo("password2", message="Passwords must match"),
+        ],
+    )
+    password2 = PasswordField("Confirm Password", validators=[DataRequired()])
+    submit = SubmitField("Sign Up")
+
+
+class LoginForm(FlaskForm):
+    """User form for login"""
+
+    name = StringField("Name", validators=[DataRequired()])
+    password = PasswordField("Password", validators=[DataRequired()])
+    submit = SubmitField("Log In")
